requirements.txt
accelerate==0.27.2
diffusers==0.26.3
numpy==1.26.3
open3d==0.18.0
pillow==10.2.0
scipy==1.12.0
torch==2.2.1
transformers==4.38.1
